update firebase + instance authentification
