+++ .gitignore
@@ -68,4 +68,3 @@ node_modules/
 # Api keys
 .keysapi
 
-VincentTestFonctionsFirebase
