+++ TheiaWeb/js/script.js
@@ -119,172 +119,9 @@ var form = document.getElementById('contactForm');
     }
 //#endregion
 
-//#region Save CONTACT FORM data firestore
-function handleFormSubmit() {
-  const form = document.getElementById('contactForm');
-  form.addEventListener('submit', async (event) => {
-      event.preventDefault();
-
-      // Get input values
-      const surname = document.getElementById('surname').value;
-      const name = document.getElementById('name').value;
-      const email = document.getElementById('email').value;
-      const phone = document.getElementById('phone').value;
-      const message = document.getElementById('message').value;
-      const now = new Date();
-      const timestamp = now.toString();
-    
-      // Gather selected services using checkboxes
-      const services = Array.from(document.querySelectorAll('.contact__choose-services-check input[type="checkbox"]:checked'))
-          .map(checkbox => checkbox.value);
-
-      // Get checkbox values
-      const cguAccepted = document.getElementById('cgu-check').checked;
-      const newsletterSubscribed = document.getElementById('newsletter-check').checked;
-
-      // Create data object
-      const userData = {
-        personalInfo: {
-          surname: surname,
-          name: name,
-          email: email,
-          phone: phone,
-          message: message
-      },
-      services: services,
-      preferences: {
-          cguAccepted: cguAccepted,
-          newsletterSubscribed: newsletterSubscribed
-      },
-      timestamp: timestamp,
-      };
-
-      // Save data to Firestore
-      const db = firebase.firestore();
-      try {
-          await db.collection('contacts').add(userData);
-          console.log('Data saved successfully.');
-          form.reset(); // Optional: Reset the form after submission          
-      } catch (error) {
-          console.error('Error saving data:', error);
-      }
-  });
-}
-
-// Call the function when the document is loaded
-document.addEventListener('DOMContentLoaded', handleFormSubmit);
-
-//#endregion
-
-//#region PopUp RGPD
-const popup = document.getElementById("popupRGPD");
-const acceptButton = document.getElementById("acceptButton");
-const refuseButton = document.getElementById("refuseButton");
-const preferencieslink = document.getElementById("popupRGPD__content-link");
-const preferencies = document.getElementById("pref-rgpd");
-const preferenciesBackBtn = document.getElementById("pref__content-btn-return");
-const cookieModal = document.getElementById("cookieModal");
-const modalAcceptButton = document.getElementById("modalAcceptButton");
-
-// Check if user has already accepted or refused cookies before
-const cookieConsent = localStorage.getItem("cookieConsent");
-
-// Function to handle cookie consent
-function handleCookieConsent(consent) {
-  localStorage.setItem("cookieConsent", consent);
-  cookieModal.style.display = "none";
-}
-
-// Event listener for accepting cookies from the modal
-modalAcceptButton.addEventListener("click", () => {
-  handleCookieConsent("accepted");
-  cookieModal.style.display = "none";
-});
-
-// Event listener for refusing cookies
-refuseButton.addEventListener("click", () => {
-  cookieModal.style.display = "block";
-});
-
-acceptButton.addEventListener("click", async () => {
-  const necessaryState = true;
-  const statisticState = true;
-  const preferencesState = true;
-  const marketingState = true;
-
-  const now = new Date();
-  const timestamp = now.toString();
-
-  const docName = "Date:" + timestamp;
-  
-  try {
-    await firestore.collection("CGU_Acceptation").doc(docName).set({
-      necessary: necessaryState,
-      statistic: statisticState,
-      preferences: preferencesState,
-      marketing: marketingState,
-      timestamp: timestamp,
-    });
-    console.log("Document successfully written!");
-    console.log("Preferences saved to Firestore!");
-  } catch (error) {
-    console.error("Error writing document: ", error);
-    console.log("An error occurred while saving preferences.");
-  }
-
-  handleCookieConsent("accepted");
-  popup.style.display = "none";
-});
-
-// Event listener for showing the preferences
-preferencieslink.addEventListener("click", () => {
-  preferencies.style.display = "flex";
-});
-
-preferenciesBackBtn.addEventListener("click", () => {
-  preferencies.style.display = "none";
-});
-
-// Get references to the switch elements
-const necessarySwitch = document.querySelector('#necessary-switch');
-const statisticSwitch = document.querySelector('#statistic-switch');
-const preferencesSwitch = document.querySelector('#preferences-switch');
-const marketingSwitch = document.querySelector('#marketing-switch');
-const acceptButtonMultiple = document.querySelector('.pref__content-btn-accept');
-
-acceptButtonMultiple.addEventListener('click', async () => {
-  const necessaryState = necessarySwitch.checked;
-  const statisticState = statisticSwitch.checked;
-  const preferencesState = preferencesSwitch.checked;
-  const marketingState = marketingSwitch.checked;
-
-  const now = new Date();
-  const timestamp = now.toString();
-  const docName = "Date: " + timestamp;
-
-  try {
-    await firestore.collection("CGU_Acceptation").doc(docName).set({
-      necessary: necessaryState,
-      statistic: statisticState,
-      preferences: preferencesState,
-      marketing: marketingState,
-      timestamp: timestamp
-    });
-    console.log("Document successfully written!");
-    console.log('Preferences saved to Firestore!');
-  } catch (error) {
-    console.error("Error writing document: ", error);
-    console.log('An error occurred while saving preferences.');
-  }
-
-  preferencies.style.display = "none";
-  popup.style.display = "none";
-});
-//#endregion
-
 //#region Player Twitch
 const clientId = 'o5x6ltrt4jwhb6ybxm0kkpjcip5mk4';
-const accessToken = 'krnc4iwqx3gl1e8c9s5d6fkamjs7ig';
+const accessToken = 'nyy3iengngzc6ghg6nmldtkmxgqaab';
 const channelId = 'theiaweb';
 
 var playerWidth = 1472;
@@ -293,31 +130,11 @@ var twitchPlayer = null; // Variable to store the Twitch player instance
 // const nextScheduledStreamChannelId = 'next_channel_id'; // Replace with the channel ID of the channel with the next scheduled stream
 // let nextStreamDateValue = "Date: Not available";
 
+
 // Call the function on page load
 window.onload = function () {
   initializePlayer();
-  fetch(`https://api.twitch.tv/helix/subscriptions?broadcaster_id=${channelId}`, {
-    headers: {
-      'Client-ID': clientId,
-      'Authorization': `Bearer ${accessToken}`
-    }
-  })
-  .then(response => {
-    if (!response.ok) {
-      throw new Error(`HTTP error! Status: ${response.status}`);
-    }
-    return response.json();
-  })
-  .then(data => {
-    // Process the data here
-    console.log(data);
-  })
-  .catch(error => {
-    console.error('Fetch error:', error);
-  });
-  
 };
-
 // Call the function whenever the window is resized
 window.addEventListener('resize', initializePlayer);
 
@@ -430,74 +247,3 @@ function initializePlayer() {
 
 //#endregion
 
-//#region Newsletter 
-const newsletterForm = document.getElementById("newsletterForm");
-const emailInput = document.getElementById("emailInput");
-const newsletterModal = document.getElementById("newsletterModal");
-
-newsletterForm.addEventListener("submit", (event) => {
-  event.preventDefault();
-
-  const email = emailInput.value;
-  const checkEmailFunctionURL = "https://us-central1-formtheia.cloudfunctions.net/checkIfEmailExists";
-
-  // Make an HTTP request to the Cloud Function to check if the email exists
-  fetch(`${checkEmailFunctionURL}?email=${encodeURIComponent(email)}`)
-    .then((response) => response.json())
-    .then((data) => {
-      if (data.exists) {
-        // The email already exists in the database
-        newsletterModal.textContent = "Vous êtes déjà inscrits a notre newsletter !";
-        newsletterModal.classList.add("active");
-        setTimeout(() => {
-          newsletterModal.classList.remove("active");
-        }, 2000);
-        emailInput.value = "";
-      } else {
-        // Continue with saving the email to the database
-        const date = new Date().toISOString().slice(0, 10);
-        const firstLetter = email.charAt(0).toLowerCase();
-        const sanitizedEmail = email.replace(/[.$#\[\]@]/g, '');
-        const customKey = `${date}_${firstLetter}_${sanitizedEmail}`;
-        newsletterModal.textContent = "Merci pour votre inscription !";
-
-        // Save the email to Firebase with the custom key
-        database
-          .ref("newsletterEmails")
-          .child(customKey)
-          .set({ email: email })
-          .then(() => {
-            console.log("Email saved successfully!");
-            // Show the modal
-            newsletterModal.classList.add("active");
-            
-            // Hide the modal after 2 seconds with fading effect
-            setTimeout(() => {
-              newsletterModal.classList.remove("active");
-            }, 2000);
-            emailInput.value = ""; // Clear the input after successful submission
-          })
-          .catch((error) => {
-            console.error("Error saving email:", error);
-          });
-      }
-    })
-    .catch((error) => {
-      console.error("Error checking email:", error);
-    });
-});
-
-// function getEmailServiceURL(email) {
-//   // Logic to determine the email service and return the appropriate URL
-//   // For example, you can check if the email domain contains "gmail.com" for Gmail users:
-//   if (email.includes("@gmail.com")) {
-//     return "https://mail.google.com/mail/u/0/#inbox";
-//   }
-
-//   // Add more cases for other email services if needed.
-
-//   // Return null for unsupported email services.
-//   return null;
-// }
-
-//#endregion
