+++ package.json
@@ -9,10 +9,10 @@
     "@google-cloud/storage": "^6.11.0",
     "animate.css": "^4.1.1",
     "cors": "^2.8.5",
-    "firebase": "^10.3.0",
+    "firebase": "^10.3.1",
     "firebase-admin": "^11.10.1",
     "firebase-functions": "^4.4.1",
-    "firebase-tools": "^12.4.7",
+    "firebase-tools": "^12.5.2",
     "jspdf": "^2.5.1",
     "nodemailer": "^6.9.4",
     "pdf-lib": "^1.17.1",
